AtividadeAvaliativa/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
mysql-connector-python==8.0.26
flask-cors==3.0.10
requests==2.26.0
